requirements.txt
colorama==0.4.4
colorlog==6.6.0
cupy==13.4.1
einops==0.8.1
flow_vis==0.1
hdf5storage==0.1.18
imageio==2.12.0
ipdb==0.13.13
matplotlib==3.5.0
mmcv_full==1.6.1
mmflow==0.5.2
munch==2.5.0
numpy==1.23.3
numpy==1.20.2
opencv_python==4.5.2.54
Pillow==8.4.0
Pillow==9.0.1
Pillow==11.2.1
PyYAML==5.4.1
PyYAML==6.0.2
qqdm==0.0.7
readchar==3.0.4
scikit_image==0.18.3
scipy==1.6.3
thop==0.1.1.post2209072238
timm==1.0.16
torch_dwconv==0.1.0
torchlights==0.3.7
torchvision==0.12.0
tqdm==4.51.0
